A bit more rust like handling of UnconfirmedServiceRequest

+++ examples/basic/simple_device.rs
@@ -88,7 +88,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // Unconfirmed request (Who-Is)
     let whois_apdu = Apdu::UnconfirmedRequest {
-        service_choice: UnconfirmedServiceChoice::WhoIs as u8,
+        service_choice: UnconfirmedServiceChoice::WhoIs,
         service_data: whois_buffer,
     };
     
@@ -98,7 +98,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let decoded_apdu = Apdu::decode(&encoded_apdu)?;
     match decoded_apdu {
         Apdu::UnconfirmedRequest { service_choice, .. } => {
-            println!("Decoded APDU service choice: {}", service_choice);
+            println!("Decoded APDU service choice: {}", service_choice as u8);
         }
         _ => println!("Unexpected APDU type"),
     }
